chore: removed inventory and added pokedex
